Change to cmake structuring. Now generate the ios project from within the ios directory. No more root Cmakelists file. This means we don't have to branch in a huge file. Also, setting the toolchain variable in the platform CMakeLists file, so we don't have to pass it into the command, making the generation a little easier to remember. Couple of little fixups since the merge with master. Namely adding new resources, removing old ones and patching up an implicit cast warning. Needed to add the security framework to the project, too. Removed all ios branching from platform CMakeLists. Bail early if we're not iOS. Buddy: Scott M

+++ src/MobileExampleApp.cpp
@@ -158,7 +158,7 @@ namespace ExampleApp
                     size_t materialsCount(materials.Size());
                     for(size_t i = 0; i < materialsCount; ++ i)
                     {
-                        const rapidjson::Value& materialJson(materials[i]);
+                        const rapidjson::Value& materialJson(materials[static_cast<int>(i)]);
                         Eegeo::Resources::Interiors::Materials::InteriorsMaterialDto dto(Eegeo::Resources::Interiors::Materials::ParseMaterial(materialJson));
                         Eegeo_ASSERT(!interiorsMaterialDtoRepository.Contains(interiorName, dto.materialName));
                         interiorsMaterialDtoRepository.Add(interiorName, dto);
